Fix: Stop canvases from getting cut from gym roof and floor
Reduce te planar distance of the calibration and workout canvases so
they don't get chopped by the player looking up at the roof or down at
the floor.
Reducing planar distance simply makes the canvases closer to thecamera.
However, the size of the UI on the canvas is unaffected.

Planar distance on both canvases was reduced from 2.0 to 1.0.

Recording of test:
https://user-images.githubusercontent.com/23128900/114325060-aae81900-9af3-11eb-877b-83c88a6c176f.mp4