Tracking: Improve Workout Screen UI for Situp
Add a progress indicator for each rep in the Workout UI.
This indicator is a stick figure showing you your rep position and
percentage complete.

A timer with the total elapsed time is also displayed.

Also fix reflections on mobile and bug where tracking didn't work with
angles close to the horizon.