Improvement: Improve Twist Crunch Tracking
Improve Twist Crunch tracking by checking for the player's X and Y
angles independently rather than dependently.

Explained:
Previously, for the app to detect the player in position 2 or 3 of a
twist crunch, the player's x and y angles had to both land within the range of a
coordinate point (x+-threshhold, y+-threshhold). This meant the player
had to turn to a very specific and narrow point in 3D space for the
system to detect their positions 2 and 3 of a twist crunch. If the
player overextended a twist crunch, the system wouldn't validate it as a
valid twist crunch.

Now, the app only checks whether the player's x and y angles are greater
than/less than a specific boundary. This means that if the player
overextends a twist crunch past his/her calibrated angles, the system
will validate the angles as a valid twist crunch. This is a much more
natural way of using the app to workout.